Fix denormalized PK inconsistency via mapping table
Crashes undergo geocoding (Port Authority + empty municipality fixes) that
updates cc/mc, but V/D/O/P retain original cc/mc from raw data. This caused
~80k orphaned vehicle records in 2023.

Solution:
- rawdata/pqt.py: Preserve original cc/mc as cc0/mc0 before geocoding
- crashes.py: Export crash_pk_mappings.parquet with (year,cc0,mc0,case)→(cc,mc)
- V/D/O/P: Apply mapping to update denormalized PKs before joining to crashes

All tables now have zero orphaned records. cc0/mc0 use Int8 (nullable) dtype.

+++ njdot/occupants.py
@@ -160,8 +160,10 @@ def map_df(df, fix_missing_vid: bool = True, drop: bool = True):
 
         # Update cc/mc where mapping exists
         # For rows without mapping, cc/mc will be NaN, so fill with original values
+        # Note: mc may be float64 (combined codes like 9901.0 for Port Authority)
         df['cc'] = df_with_mapping['cc'].fillna(df['cc']).astype('int8')
-        df['mc'] = df_with_mapping['mc'].fillna(df['mc']).astype('int8')
+        # Convert mc to float to handle combined codes from crashes
+        df['mc'] = df_with_mapping['mc'].fillna(df['mc'].astype('float64'))
 
         num_updated = df_with_mapping['cc'].notna().sum()
         err(f"  Updated {num_updated:,} occupant PKs from mapping table")
